chore(clerk-js): Move shared utils to common
This utility was bundled with the OrganizationProfile for no reason
