+++ packages/clerk-js/src/ui/components/SignUp/util.ts
@@ -1,42 +1,1 @@
-import type { SignUpResource } from '@clerk/types';
-
-type CompleteSignUpFlowProps = {
-  signUp: SignUpResource;
-  verifyEmailPath?: string;
-  verifyPhonePath?: string;
-  navigate: (to: string) => Promise<unknown>;
-  handleComplete?: () => Promise<void>;
-  redirectUrl?: string;
-  redirectUrlComplete?: string;
-};
-
-export const completeSignUpFlow = ({
-  signUp,
-  verifyEmailPath,
-  verifyPhonePath,
-  navigate,
-  handleComplete,
-  redirectUrl = '',
-  redirectUrlComplete = '',
-}: CompleteSignUpFlowProps): Promise<unknown> | undefined => {
-  if (signUp.status === 'complete') {
-    return handleComplete && handleComplete();
-  } else if (signUp.status === 'missing_requirements') {
-    if (signUp.missingFields.some(mf => mf === 'saml')) {
-      return signUp.authenticateWithRedirect({
-        strategy: 'saml',
-        redirectUrl,
-        redirectUrlComplete,
-        continueSignUp: true,
-      });
-    }
-
-    if (signUp.unverifiedFields?.includes('email_address') && verifyEmailPath) {
-      return navigate(verifyEmailPath);
-    }
-    if (signUp.unverifiedFields?.includes('phone_number') && verifyPhonePath) {
-      return navigate(verifyPhonePath);
-    }
-  }
-  return;
-};
+export * from '../../utils/completeSignUpFlow';
