+++ packages/clerk-js/src/ui/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './completeSignUpFlow';
 export * from './fromEntries';
 export * from './containsAllOf';
 export * from './createInfiniteAccessProxy';
